+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fix using `host` header instead of `:authority`
 * Make domain regex matches even more strict: limit amount of subdomains
 * Automatically remove problematic domains from Greenlock (though it does not work)
+* Keep track of IP reputations
 
 ## 0.5.3 (2023-10-19)